Create [Robot-Framework] How to check login page.
You have to change locator in it. 
In this page, When I login fail => can not navigate to new url such as http://58.137.96.185/#/login/fail and opposite, so I have to based on Title of page. 
If your page has a url/fail. You just need to replace ${Dashborad} = {$url} url/fail and you have to add more url/successfully to popup for test-case login susseccfully. Use another key-work such as Location Should Be , not Title should be.